Integrates Rust core build and fixes CRC32 return type
- Automates building and copying of the native Rust library as part of the .NET project build process, ensuring cross-platform support.
- Updates the CRC32 interop signature to use the correct return type, preventing potential interop issues and data truncation.

+++ src/Core/Interop/RustCrypto.cs
@@ -75,7 +75,7 @@ public static class RustCrypto
     /// <param name="len">The length of the data.</param>
     /// <returns>The CRC32 checksum as an unsigned integer.</returns>
     [DllImport(DllName, CallingConvention = CallingConvention.Cdecl)]
-    private static extern unsafe nuint ffi_calculate_crc32(byte* data, nuint len);
+    private static extern unsafe uint ffi_calculate_crc32(byte* data, nuint len);
 
     // --- SAFE WRAPPERS ---
